Normalize all the line endings
Used the Windows Github GUI to add a .gitattributes file specifying
which files are actually text. This commit is to normalize all line
endings for text files. This is how I think it should be. (The previous
way was also a pain as on Windows my Stata editor would always switch
around the line endings and making commits without line ending changes
was time consuming).