use csh-friendly stderr redirection
This fixes parallel for (t)csh users. This syntax is compatible with
at least bash, zsh and tcsh, which should cover the majority of
Linux and MacOS users.

Users of ksh and plainer sh varieties will need to override their
shell to one of the above prior to running Stata, e.g.
    env SHELL=/bin/bash xstata ...